refactor: move in-line script to JS file and split chatView/utils into separate files for better manageability
