+++ resources/chat-view/chat.css
